Changed the way string escapes are handled, to dramatically reduce allocation
Instead of eagerly bailing out to a heap allocation whenever an escape character is encountered, we simply mark the string slice as a candidate for escaping. In most cases this never needs to be done; JSON output needs to be re-escaped anyway, and only a handful of operations actually care about treating escaped characters as what they really represent. So dynamic string creation is done lazily, as required, and in most cases it never happens at all.